+++ frontend/src/components/NavBar/Teacher.jsx
@@ -5,13 +5,37 @@ import {
   FiSettings,
   FiHelpCircle,
   FiUser,
+  FiLogOut,
 } from 'react-icons/fi';
+import {
+  selectCurrentUser,
+  selectCurrentToken,
+  logOut,
+} from "../../features/auth/authSlice";
+import { useDispatch } from "react-redux";
+import { useNavigate } from "react-router-dom";
+import { persistor } from "../../store";
 
 export const Teacher_NavBar = () => {
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
   const [isNavbarCollapsed, setIsNavbarCollapsed] = useState(true);
 
   const handleNavbarToggle = () => {
     setIsNavbarCollapsed(!isNavbarCollapsed);
+  };
+
+  
+  const handleLogout = async (e) => {
+    e.preventDefault();
+
+    try {
+      dispatch(logOut);
+      persistor.purge();
+      navigate("/");
+    } catch (error) {
+      console.log(error);
+    }
   };
 
   return (
@@ -37,13 +61,10 @@ export const Teacher_NavBar = () => {
               {isNavbarCollapsed ? <FiBell /> : 'Notification'}
             </span>
             <span className="nav-item nav-link">
-              {isNavbarCollapsed ? <FiSettings /> : 'Settings'}
-            </span>
-            <span className="nav-item nav-link">
-              {isNavbarCollapsed ? <FiHelpCircle /> : 'Help'}
-            </span>
-            <span className="nav-item nav-link">
               {isNavbarCollapsed ? <FiUser /> : 'User'}
+            </span>
+            <span className="nav-item nav-link"  onClick={handleLogout}>
+              {isNavbarCollapsed ? <FiLogOut /> : 'Logout'}
             </span>
           </div>
         </div>
